backend/requirements.txt
fastapi
uvicorn[standard]
sumy
gensim
transformers
PyMuPDF
python-docx
newspaper3k
pytesseract
Pillow
requests
beautifulsoup4
